requirements.txt
torch>=1.8.0
torchvision>=0.9.0
matplotlib>=3.3.4
numpy>=1.19.5
scikit-learn>=0.24.1
pillow>=8.1.0
tqdm>=4.56.0
pandas>=1.2.1